retrieval_service/requirements.txt
asyncpg==0.30.0
fastapi==0.115.0
google-auth==2.35.0
google-cloud-firestore==2.19.0
google-cloud-aiplatform==1.72.0
google-cloud-spanner==3.49.1
langchain-core==0.3.18
pgvector==0.3.5
pydantic==2.9.0
uvicorn[standard]==0.31.0
cloud-sql-python-connector==1.12.1
google-cloud-alloydb-connector[asyncpg]==1.4.0
sqlalchemy[asyncio]==2.0.36
pandas==2.2.3
pandas-stubs==2.2.2.240807
langchain-text-splitters==0.3.0
langchain-google-vertexai==2.0.7
asyncio==3.4.3
datetime==5.5
pymysql==1.1.1
types-PyMySQL==1.1.0.20240524
neo4j==5.26.0
sqlparse==0.5.2
